requirements.txt
# ===== Core: Telegram listener =====
# telethon==1.34.0
# aiosqlite>=0.19,<1.0
# python-dotenv>=1.0,<2.0

# ===== Web / ASGI =====
# Flask[async]>=3.0,<4.0
# hypercorn>=0.16,<1.0

# --- Web Server & ASGI ---
# Flask 3.0+ поддерживает асинхронные маршруты
Flask>=3.0.0
# Hypercorn нужен для запуска асинхронного цикла (Flask + Telethon)
hypercorn>=0.16.0

# --- Database & ORM ---
# Основной движок для работы с БД
SQLAlchemy>=2.0.0
# Асинхронный драйвер для SQLite
aiosqlite>=0.19.0

# --- Telegram bot ---
pyTelegramBotAPI>=4.14,<5.0
requests>=2.32,<3.0

# --- Telegram API ---
# Работа с Telegram через MTProto (намного мощнее обычных ботов)
telethon>=1.34.0

# --- Environment & Utils ---
# Для хранения API_ID, API_HASH и путей в .env файле
python-dotenv>=1.0.0
# Утилиты для путей (обычно встроены, но полезно для окружения)
pathlib>=1.0.1

# ===== Utils =====
urllib3
chardet